add tests for import command
